+++ tests/bootstrap.php
@@ -14,3 +14,5 @@ require_once 'vendor/autoload.php';
 
 define('DS', DIRECTORY_SEPARATOR);
 define('ROOT', dirname(__DIR__) . DS);
+define('COMPARING_FILES', ROOT . 'tests' . DS . 'comparing_files' . DS);
+define('TMP', sys_get_temp_dir() . DS . 'php-tools' . DS);